use shared schema for hstore and apartment
So, turns out that apartment doesn't support structure.sql (yet)
(influitive/apartment/pull/16), so we must go back to schema.rb
The only way to make apartment work with hstore is to have a persistent
schema, that has the hstore extension installed. I've added this step to
the installation procedure. It is not optimal, but makes everything
work.